+++ packages/design-system/src/scalars/components/date-picker-field/date-picker-validations.ts
@@ -1,13 +1,12 @@
 import { format } from "date-fns";
 import { DatePickerFieldProps } from "./date-picker-field";
-import {
-  formatDateToValidCalendarDateFormat,
-  getDateFormat,
-  getDateFromValue,
-  normalizeMonthFormat,
-} from "./utils";
+import { formatDateToValidCalendarDateFormat, getDateFromValue } from "./utils";
 import { DateFieldValue } from "./types";
-import { isDateFormatAllowed } from "../date-time-field/utils";
+import {
+  getDateFormat,
+  isDateFormatAllowed,
+  normalizeMonthFormat,
+} from "../date-time-field/utils";
 
 export const validateDatePicker =
   ({ minDate, maxDate, dateFormat }: DatePickerFieldProps) =>
