+++ packages/design-system/src/scalars/components/date-time-field/date-time-field-validations.ts
@@ -1,13 +1,16 @@
 import { format } from "date-fns";
 import {
   formatDateToValidCalendarDateFormat,
-  getDateFormat,
   getDateFromValue,
-  normalizeMonthFormat,
   splitIso8601DateTime,
 } from "../date-picker-field/utils";
 import { DatePickerFieldProps } from "../date-picker-field/date-picker-field";
-import { isDateFormatAllowed, isValidTime } from "./utils";
+import {
+  getDateFormat,
+  isDateFormatAllowed,
+  isValidTime,
+  normalizeMonthFormat,
+} from "./utils";
 import { DateFieldValue } from "../date-picker-field/types";
 
 export const dateTimeFieldValidations =
